Restructured to have card updates and valid and invalid snooze times
A root card is displayed for the first runthrough for onGmailMessage, then each time a new snooze time is selected, it will update the card. In order to do that, the snooze widgets have been pulled out as their own functions, and there is now and original onGmailMessage function and an updateCard function to update this card build.